Simple Arithmetic Calculator
This repository features a simple calculator implemented in C++. The calculator is capable of performing four fundamental arithmetic operations: addition (+), subtraction (-), multiplication (*), and division (/). The user can input operands and operators to execute these operations, facilitating basic mathematical computations.
Furthermore, this calculator includes a feature that allows the user to perform an operation again on the previous answer, enabling a seamless and efficient calculation process.
This code serves as a helpful educational resource for C++ learners and enthusiasts, offering insight into the implementation of fundamental arithmetic operations within a programming context. You are welcome to explore the code, contribute to its development, or use it as a reference for similar projects. Leveraging GitHub's collaborative features and version control, this repository provides an ideal platform for learning, sharing, and enhancing the functionality of the calculator.

#codsoft