To-Do List Program in C++
This repository houses a simple to-do list program created as part of the 4th task during my internship at CodSoft. The program, written in C++, is designed to be used in the terminal environment. It offers users the capability to manage their tasks effectively with four essential features:

1. Adding new tasks to the list
2. Viewing the entire task list
3. Marking tasks as done
4. Removing tasks from the list

This project serves as a practical showcase of my skills in C++ programming and my ability to develop user-friendly command line applications. The program is not open for contributions or changes at this time. It is intended for demonstration purposes only.

Please feel free to explore the repository. Any questions or feedback are greatly appreciated.
#codsoft